pyproject.toml
[tool.poetry]
name = "cheesegpt"
version = "0.1.0"
description = "an rag example"
authors = ["alexjacobs <alex.jacobs08@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
wikipedia-api = "^0.6.0"
openai = "0.28.1"
langchain = "^0.0.339"
py-dotenv = "^0.1"
python-dotenv = "^1.0.0"
redis = "^5.0.1"
tiktoken = "^0.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
